Factored out generators and shrinkers
